basic working condition

+++ automation.py
@@ -1,16 +1,4 @@
-import pyautogui, time
-from pyautogui import keyDown, keyUp, press
-import os, path
-
-pyautogui.FAILSAFE = True
-here = path.abspath(path.dirname(__file__))
-
-def click(image, timeout=None, double=False):
-    l = None
-    start = time.time()
-    while l is None:
-        l = pyautogui.locateOnScreen(here + "/images/" + image)
-        elapsed = time.time() - startimport pyautogui, time
+import pyautogui, time, subprocess
 from pyautogui import keyDown, keyUp, press
 import socket, threading
 from threading import Thread
@@ -32,14 +20,16 @@ here = path.abspath(path.dirname(__file__))
 rcvd = ""
 
 def click(image, timeout=None, double=False, alt=False):
+    print("click({}, timeout={}, double={}, alt={})".format(image,timeout,double,alt))
     l = None
     start = time.time()
     while l is None:
-        l = pyautogui.locateOnScreen(here + "/images/" + image + ".PNG")
+        l = pyautogui.locateOnScreen(here + "\\images\\" + image + ".PNG")
         elapsed = time.time() - start
         if alt and l is None:
-            l = pyautogui.locateOnScreen(here + "/images/" + image + "_alt.PNG")
+            l = pyautogui.locateOnScreen(here + "\\images\\" + image + "_alt.PNG")
         if timeout is not None and elapsed >= timeout:
+            print("click: timeout: elapsed: {}".format(elapsed))
             break
     ret = l is not None
     if ret:
@@ -55,21 +45,44 @@ def rx(sock):
         time.sleep(0.2)
 
 
-def restart():
-    os.system("taskkill /f /im  JProg_7361.exe")
-    keyDown("winleft"); press("d"); keyUp("winleft")
-    click("showdesk", timeout=10, double=False)
-    click("folder", timeout=10, double=True)
-    click("exe", timeout=10, double=True, alt=True)
-    click("runanyway", timeout=10, double=False)
-    click("interface", timeout=10, double=False)
-    click("cleartoship", timeout=10, double=False)
+def restart(kill=True):
+    print("restarting")
+    if rcvd == "pause":
+            return
+    if kill:
+        os.system("taskkill /f /im  JProg_7361.exe")
+        if rcvd == "pause":
+            return
+    #keyDown("winleft"); press("d"); keyUp("winleft")
+    #if rcvd == "pause":
+     #       return
 
+    subprocess.Popen(["C:/Users/model/Desktop/7361-NJ-Nissan_Kiks/JProg_7361.exe"])
 
+    for t in [("interface", 10, False, True),("cleartoship", 10, False, False)]:
+        print("trying to click {}".format(t[0]))
+        if rcvd == "pause":
+            print("pause")
+            return
+        click(t[0], timeout=t[1], double=t[2], alt=t[3])
+    
 def main(sock):
     global rcvd
-    while 1: 
-        for i, step in enumerate([("start",10),("no",10),("proceed",10),("c_ok",20),("proceed",10),("e_ok",20)]):
+    restart(kill=False)
+    while 1:
+        while rcvd == "pause":
+            time.sleep(0.1)
+        for i, step in enumerate([
+            ("start",10, False),
+            ("no",10, False),
+            ("proceed",10, False),
+            ("c_ok",20, True),
+            ("proceed",10, False),
+            ("e_ok",20, False)
+        ]):
+            if rcvd == "pause":
+                break
+            print(step[0])
             if i == 2:
                 sock.sendto(bytes("proceed", "utf-8"), address)
                 for i in range(50):
@@ -77,45 +90,16 @@ def main(sock):
                         break
                     time.sleep(0.1)
                 rcvd = "" if rcvd == "ok" else "restart"
-            if rcvd == "restart" or not click(step[0], timeout=step[1]):
+            result = click(step[0], timeout=step[1], alt=step[2])
+            print("click: {}: result: {}".format(step[0], result))
+            if rcvd == "restart" or not result:
                 restart()
                 break
 
 if __name__ == "__main__":
-    address = ("192.168.7.2", 11005)
+    address = ("192.168.7.2", 11000)
     sock = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP)
-    sock.settimeout(5)
     sock.bind(("", 11000))
     rxt = Thread(target=rx,args=([sock]))
     rxt.start()
     main(sock)
-
-
-
-        if timeout is not None and elapsed >= timeout:
-            break
-    ret = l is not None
-    if ret:
-        for i in range(double + 1):
-            pyautogui.click(l[0] + 5, l[1] + 10)
-    return ret
-
-def restart():
-    os.system("taskkill /f /im  JProg_7361.exe")
-    keyDown("winleft"); press("d"); keyUp("winleft")
-    click("jprog_icon.PNG", timeout=10, double=True)
-    click("interface.PNG", timeout=10, double=True)
-    click("screw.PNG", timeout=10, double=True)
-    click("arrow.PNG", timeout=10, double=True)
-    click("cleartoship.PNG", timeout=10, double=True)
-
-def main():
-    while 1: 
-        for step in [("start",10),("no",10),("proceed",10),("c_ok",20),("proceed",10),("e_ok",20)]:
-            if not click(step[0] + ".PNG", timeout=step[1]):
-                restart()
-
-if __name__ == "__main__":
-    main()
-
-
